+++ games-emulation/common-shaders/common-shaders-9999-r2.ebuild
@@ -1,4 +1,4 @@
-# Copyright 1999-2016 Gentoo Foundation
+# Copyright 1999-2017 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
 # $Id$
 
@@ -8,16 +8,7 @@ inherit libretro
 
 DESCRIPTION="Collection of commonly used Cg shaders."
 HOMEPAGE="https://github.com/libretro/common-shaders"
-SRC_URI=""
-
-if [[ ${PV} == 9999 ]]; then
-	inherit git-r3
-
-	EGIT_REPO_URI="https://github.com/libretro/common-shaders.git"
-	KEYWORDS=""
-else
-	KEYWORDS="~amd64 ~x86"
-fi
+KEYWORDS=""
 
 LICENSE="GPL-3"
 SLOT="0"
@@ -31,7 +22,7 @@ DEPEND="${RDEPEND}
 src_install() {
 	dodir "${LIBRETRO_DATA_DIR}"/shaders
 	# Remove unnecessary git files
-	rm -r .git
+	[[ ! ${PV} == "1.0_pre"* ]] && rm -r .git
 	# Install plain CG shaders
 	use cg && cp -R "${S}"/* "${D}${LIBRETRO_DATA_DIR}"/shaders/
 	if ! use cg; then
